Support printing of current line number
This commit adds support for '=' and 'l'.
'=' prints the current line number, just like it does in 'more'.
'l' is an extra tool that prints the line number before every line
from then on (until you press 'l' again).
This is similar to '-N' in less (but I couldn't use 'n' as that is used to repeat a search).